Integrate Location and MapBox
